# Conflicts:
#	src/main/java/ru/skypro/homework/service/ImageService.java
#	src/main/java/ru/skypro/homework/service/impl/ImageServiceImpl.java

+++ src/main/java/ru/skypro/homework/dto/CommentDto.java
@@ -9,11 +9,8 @@ public class CommentDto {
     private Integer author;
     private String authorImage;
     private String authorFirstName;
-    private Integer createdAt;
+    private long createdAt;
     private Integer pk;
     private String text;
-
-
-
 
 }
